Removing single quote from markdown syntax table.
This was causing issues with markdown files where apostrophes would be
considered the start of a string delimiter, and would thus be colored as
though they were a string. While removing the single quote from the
syntax table will prevent expand-region from being able to select inside
single-quoted strings, this shouldn't be much of an issue. The tradeoff
for not making Markdown files look weird should be well worth it.