helper function to update emacs-startup-directory
After removing the functionality, in the last commit, to pin the
default-directory to the directory in which we intially opened Emacs, I
figured there may be some circumstances in which you would want to
update the emacs-startup-directory.

For exmaple, the MikeFuzzyFileFinder function searches recursively in
emacs-startup-directory, and you may not want the directory in which you
initially opened Emacs to be the one in which the recursive search starts.