Removing single-quote as delimiter for YAML.
In lines with the Ansible "name:" item, the single-quote would be
interpreted as the start of a string. As a result, if you had a name
line with an apostrophe in it (like "this isn't going to work"), you'd
have all of the text in the file highlighted as a string until Emacs
encountered the next apostrophe.