Finally figured out why tramp-mode was soo slow!
The problem here was projectile. It constantly tries to figure out what
project name to display for the Emacs modeline, and this causes
everything (including editing an already-opened remote file) to be
insanely sluggish! By setting the mode line to be a string (I simply
chose an empty string), we prevent Projectile from trying to
programmatically determine it.