Use isearch mode mapping for highlighting, instead of isearch end hook
This solves the problem of not being able to press M-% during an active iSearch to activate isearch-query-replace-regexp. Entering isearch-query-replace-regexp causes the isearch end hook to be called, which selects the search term as the active region. We really only want this behavior when hitting <return> during an isearch. Removing the hook and adding an isearch mode mapping gives us the desired behavior.