Switch to read-char so that we are not prompted for each cursor
The multiple cursors library will use the result from a single read-char call,
and repeat that input for all cursors.

Previously, when using a function argument via interactive, we were being
prompted for the character to expand to for each cursor. By switching to
read-char, we are prompted once for a character, and then all cursors expand to
that character without any further user input.